Remove unused hoc uses and just use it in the one place
