+++ frontend/src/metabase/collections/components/CollectionCopyEntityModal.tsx
@@ -4,7 +4,6 @@ import { t } from "ttag";
 
 import Collections from "metabase/entities/collections";
 import EntityCopyModal from "metabase/entities/containers/EntityCopyModal";
-import withToast from "metabase/hoc/Toast";
 import { connect } from "metabase/lib/redux";
 import { entityTypeForObject } from "metabase/lib/schema";
 
@@ -69,4 +68,4 @@ function CollectionCopyEntityModal({
 }
 
 // eslint-disable-next-line import/no-default-export
-export default withToast(connect(mapStateToProps)(CollectionCopyEntityModal));
+export default connect(mapStateToProps)(CollectionCopyEntityModal);
